toy_project/pyproject.toml
[tool.poetry]
name = "toy-project"
version = "0.1.0"
description = ""
authors = ["akinoritoma"]
readme = "README.md"
packages = [{include = "toy_project"}]

[tool.poetry.dependencies]
python = "^3.8"
djoser = "^2.2.0"
djangorestframework = "^3.14.0"
django = "^4.2.1"
environ = "^1.0"
django-phone-field = "^1.8.1"
django-cors-headers = "^4.0.0"
django-cleanup = "^7.0.0"
django-filter = "^23.2"
django-rest-passwordreset = "^1.3.0"
mysqlclient = "^2.1.1"
djangorestframework-simplejwt = "^5.2.2"
pillow = "^9.5.0"
django-environ = "^0.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
